코드스타일 수정 및 InspectSignUpForm에서 프로토콜과 클래스파일 분리
