+++ SignUp/SignUp/Model/InspectPassword.swift
@@ -1,46 +1,11 @@
 //
-//  Validate.swift
+//  InspectPassword.swift
 //  SignUp
 //
-//  Created by 심영민 on 2021/03/30.
+//  Created by 심영민 on 2021/04/02.
 //
 
-import UIKit
-
-@objc protocol Inspectable {
-    @objc optional func inspect(with item: String) -> Any
-    @objc optional func inspect(with item: String, with item2: String) -> Any
-}
-
-class InspectID: Inspectable {
-
-    enum IDMessage: String {
-        case Range = "5자이상 20자 이하로 입력해주세요"
-        case Used = "이미 사용중인 아이디입니다"
-        case Success = "사용 가능한 아이디입니다"
-    }
-    
-    private var idList: [String] = []
-    
-    init() {
-        NetworkManager().serialization() { (list) in
-            self.idList = list
-        }
-    }
-    
-    func inspect(with item: String) -> Any{
-        let temp = NSPredicate(format: "SELF MATCHES %@", "[a-z0-9-_]{5,20}")
-        
-        if !temp.evaluate(with: item) {
-            return IDMessage.Range
-        }
-        else if idList.contains(item){
-            return IDMessage.Used
-        } else {
-            return IDMessage.Success
-        }
-    }
-}
+import Foundation
 
 class InspectPassword: Inspectable {
     
@@ -101,38 +66,5 @@ class InspectPassword: Inspectable {
             return PasswordMessage.Range
         }
         return nil
-    }
-}
-
-class InspectUserName: Inspectable {
-    
-    enum UserNameMessage: String {
-        case Empty = "이름은 필수 입력 항목입니다"
-        case Success = ""
-    }
-    
-    func inspect(with name: String) -> Any {
-        if name.isEmpty {
-            return UserNameMessage.Empty
-        }
-        else {
-            return UserNameMessage.Success
-        }
-    }
-}
-
-class InspectRecheckPassword: Inspectable {
-    
-    enum PasswordRecheckMessage: String {
-        case Same = "비밀번호가 일치합니다"
-        case Different = "비밀번호가 일치하지 않습니다"
-    }
-    
-    func inspect(with item: String, with item2: String) -> Any {
-        if item == item2 {
-            return PasswordRecheckMessage.Same
-        } else {
-            return PasswordRecheckMessage.Different
-        }
     }
 }
